+++ typepadapp/__init__.py
@@ -33,7 +33,7 @@ typepadapp provides a basis for TypePad API applications built with Django.
 
 """
 
-__version__ = '1.1.3'
+__version__ = '1.2'
 __date__ = '20 April 2010'
 __author__ = 'Six Apart Ltd.'
 __credits__ = """Brad Choate